Modularise fn()s
